+++ src/app/ffbe/model/effects/passives/passive-dual-wield-weapon-category-unlock-effect.model.ts
@@ -1,8 +1,8 @@
-import {EffectParser} from '../../../../mappers/effects/effect-parser';
-import {Skill} from '../../../skill.model';
-import {SkillEffect} from '../../skill-effect.model';
-import {TargetNumberEnum} from '../../target-number.enum';
-import {TargetTypeEnum} from '../../target-type.enum';
+import {EffectParser} from '../../../mappers/effects/effect-parser';
+import {Skill} from '../../skill.model';
+import {SkillEffect} from '../skill-effect.model';
+import {TargetNumberEnum} from '../target-number.enum';
+import {TargetTypeEnum} from '../target-type.enum';
 
 export class PassiveDualWieldWeaponCategoryUnlockEffect extends SkillEffect {
